requirements.txt
python-telegram-bot==20.5
httpx
requests
gTTS
pydub
SpeechRecognition
